// src/app/core/models/sign-up-for-a-tour-contacted/sign-up-for-a-tour-contacted.model.ts
export class SignUpForATourContacted{
  id: string;
  tenGV: string;
  tenKhoa: string;
  nganh: string;
  soLuongSV: number;
  sinhVienNam: number;
  thoiGianDuKien: string;
  tenCongTy: string;
  diaChi: string;
  hocKi: string;
  namHoc: string;
  noiDungThamQuan: string;
  yeuCauCongVan: boolean;
}